structure changes (#10)
* not connected is better than disconnected, because disconnected could mean "I was connected but I lost my connection"

* move functionality into its own component in Connected

* moved transaction tests into connected folder

* use index.tsx directly, if you have use them at all

* simplified structure
