--- src/app/components/blockchain/ChainEnvironment/index.tsx
@@ -1,3 +1,0 @@
-import ChainEnvironment from "./ChainEnvironment";
-
-export default ChainEnvironment;
