--- src/app/components/blockchain/Disconnected/index.tsx
@@ -1,3 +1,0 @@
-import Disconnected from "./Disconnected";
-
-export default Disconnected;
