--- src/app/components/blockchain/NativeBalance/index.tsx
@@ -1,3 +1,0 @@
-import NativeBalance from "./NativeBalance";
-
-export default NativeBalance;
