+++ src/app/globals.css
@@ -111,4 +111,4 @@ button:focus-visible {
   }
 }
 
-@import "./components/FullScreenLoading/FullScreenLoading.module.css";
+@import "./components/FullScreenLoading/style.css";
